Fixing some error-handling logic in the sendMsg code.
This could deadlock, because MVars require every write to be matched
with a read. Read (or write) twice in a row and the code freezes.

I've replaced MVar with a TVar, which is more flexible, and in doing so
I've been able to tidy up the code a little too.